feat: 핀 클러스터링 기능 완성 (적용은 테스트용으로만 적용상태), 더미마커 생성 함수 추가 #43
